api/requirements.txt
fastapi
uvicorn
psycopg2
tomli
pydantic
bcrypt
PyJWT
